reorg to wallet, add TODO's
